requirements.txt
pdfplumber
pandas
requests
python-dotenv